Conflicts:
	src/main/java/com/example/restproyect/states/Destete.java
	src/main/java/com/example/restproyect/states/Mob.java
	src/main/java/com/example/restproyect/states/objetosinternos/destete/EstadoDestete.java
	src/main/java/com/example/restproyect/states/objetosinternos/destete/VariacionDestete.java
	src/main/java/com/example/restproyect/states/objetosinternos/engorde/VariacionEngorde.java
	src/main/java/com/example/restproyect/states/objetosinternos/mobs/Submob.java
	src/main/java/com/example/restproyect/states/objetosinternos/mobs/VariacionMob.java
	src/main/java/com/example/restproyect/states/objetosinternos/mobs/VariacionesMobs.java
	src/main/java/com/example/restproyect/states/objetosinternos/mobs/WeaningMobs.java

+++ src/main/java/com/example/restproyect/hilos/tareas/TareaEngorde.java
@@ -22,7 +22,7 @@ public class TareaEngorde extends AbsTarea {
 		this.numero = numero;
 		
 	}
-	private Node cambiarValoresMes(List<Integer> valores, Node nodo) {
+	private Node cambiarValoresMes(List<Float> valores, Node nodo) {
 		nodo.getAttributes().getNamedItem("January").setNodeValue(String.valueOf(valores.get(0)));
 		nodo.getAttributes().getNamedItem("February").setNodeValue(String.valueOf(valores.get(1)));
 		nodo.getAttributes().getNamedItem("March").setNodeValue(String.valueOf(valores.get(2)));
@@ -54,7 +54,7 @@ public class TareaEngorde extends AbsTarea {
 				
 				doc.setDocumento(insertDoc);
 				//Para cada tag dentro del tag <escenario> Busco los tags que tienen las variaciones
-				NodeList node = doc.getDocumento().getChildNodes().item(0).getChildNodes();		
+				NodeList node = doc.getDocumento().getChildNodes().item(0).getChildNodes();
 				for(int j=0; j < node.getLength(); j++) {
 					/*
 					 * indice par es un text dentro de los tags, solo
